\name{bhattacharyya_dist}
\alias{bhattacharyya_dist}
\docType{methods}
\title{Bhattacharyya distance for univariate Gaussian distributions.}
\description{
    Measures Bhattacharyya distance for two univatiate Gaussian
    distributions.
}
\usage{
    bhattacharyya_dist( mu1, mu2, sigma1, sigma2 )
}
\arguments{
    \item{mu1}{mean of the first Gaussian distribution.}
    \item{mu2}{mean of the second Gaussian distribution.}
    \item{sigma1}{standard deviation of the first Gaussian distribution.}
    \item{sigma2}{standard deviation of the second Gaussian distribution.}
}
\value{
    Bhattacharyya distance as double.
}
\author{Andrius Merkys}
