\name{digamma_approx}
\alias{digamma_approx}
\docType{methods}
\title{Calculate Approximate Value of The Digamma Function.}
\description{
    Calculates approximate value of the digamma function using first
    eight non-zero members of asymptotic expression for digamma(x).
    Implemented according to Wikipedia.
}
\usage{
    digamma_approx( x )
}
\arguments{
    \item{x}{data vector}
}
\value{
    Digamma function value.
}
\references{
    Users of Wikipedia.
    Digamma function.
    \url{https://en.wikipedia.org/w/index.php?title=Digamma_function&oldid=708779689}
}
\author{Andrius Merkys}
