\name{plot_circular_hist}
\alias{plot_circular_hist}
\docType{methods}
\title{Mixture Distribution Modelling}
\description{
    Plot a circular histogram.
}
\usage{
    plot_circular_hist( x, breaks = 72, ball = 0.5, ... )
}
\arguments{
    \item{x}{data vector}
    \item{breaks}{number of breaks in histogram}
    \item{ball}{radius of the drawn circle}
    \item{...}{parameters passed to plot()}
}
\author{Andrius Merkys}
