\name{wmedian}
\alias{wmedian}
\docType{methods}
\title{Calculate Weighted Median.}
\description{
    Calculated weighted median.
}
\usage{
    wmedian( x, w, start = 1, end = length( x ) )
}
\arguments{
    \item{x}{sample vector}
    \item{w}{weights vector}
    \item{start}{start index (default: 1)}
    \item{end}{end index (default: last index in \emph{x})}
}
\value{Median}
\references{
    Users of Wikipedia.
    Weighted median.
    \url{https://en.wikipedia.org/w/index.php?title=Weighted_median&oldid=690896947}
}
\author{Andrius Merkys}
